move synchronization utils to localstack.utils.sync
