+++ localstack/utils/generic/dict_utils.py
@@ -1,3 +1,2 @@
 # DEPRECATED: use localstack.utils.collections
-
 from ..collections import AccessTrackingDict, get_safe, pick_attributes, set_safe_mutable  # noqa
